+++ includes/acl/core/track_types.h
@@ -222,15 +222,15 @@ namespace acl
 	constexpr uint8_t k_invalid_bit_rate = 0xFF;
 	constexpr uint8_t k_lowest_bit_rate = 1;
 	constexpr uint8_t k_highest_bit_rate = sizeof(k_bit_rate_num_bits) - 1;
-	constexpr uint8_t k_num_bit_rates = sizeof(k_bit_rate_num_bits);
+	constexpr uint32_t k_num_bit_rates = sizeof(k_bit_rate_num_bits);
 
 	static_assert(k_num_bit_rates == 19, "Expecting 19 bit rates");
 
 	// If all tracks are variable, no need for any extra padding except at the very end of the data
 	// If our tracks are mixed variable/not variable, we need to add some padding to ensure alignment
-	constexpr uint8_t k_mixed_packing_alignment_num_bits = 16;
+	constexpr uint32_t k_mixed_packing_alignment_num_bits = 16;
 
-	inline uint8_t get_num_bits_at_bit_rate(uint8_t bit_rate)
+	inline uint32_t get_num_bits_at_bit_rate(uint8_t bit_rate)
 	{
 		ACL_ASSERT(bit_rate <= k_highest_bit_rate, "Invalid bit rate: %u", bit_rate);
 		return k_bit_rate_num_bits[bit_rate];
